Add Scene::Loader
- Remove now unused headers from scene.hpp.
- Move all loading logic into `Scene::Loader`.
- WIP: temp code with futures commented out.
